+++ components/containers/CardsColumns.js
@@ -10,33 +10,47 @@ import styles from '@/styles/CardsColumns.module.scss';
 export default function CardsColumns({ cards, customClass }) {
   return (
     <Container>
-      <Swiper
-        className={styles.swiper}
-        mousewheel={true}
-        grabCursor={true}
-        modules={[Pagination]}
-        pagination
-        centerInsufficientSlides={true}
-        slidesPerView={1}
-        breakpoints={{
-          768: {
-            slidesPerView: 2,
-            spaceBetween: 20,
-          },
-          1334: {
-            slidesPerView: cards.length,
-            spaceBetween: cards.length < 3 ? 75 : 20,
-          },
-        }}
+      <div
+        className={
+          customClass
+            ? `${styles.cardColumns} ${styles[customClass]}`
+            : styles.cardColumns
+        }
       >
-        {cards.map((card, index) => (
-          <SwiperSlide key={index} className={styles.swiperSlide}>
-            <div className={styles.inner__content}>
-              <Card key={index} card={card} customClass={customClass} />
-            </div>
-          </SwiperSlide>
-        ))}
-      </Swiper>
+        <Swiper
+          className={styles.swiper}
+          mousewheel={true}
+          grabCursor={true}
+          modules={[Pagination]}
+          pagination
+          centerInsufficientSlides={false}
+          slidesPerView={1}
+          breakpoints={{
+            768: {
+              slidesPerView: 2,
+              spaceBetween: 20,
+            },
+            1334:
+              customClass === 'blog'
+                ? {
+                    slidesPerView: 3,
+                    spaceBetween: 20,
+                  }
+                : {
+                    slidesPerView: cards.length,
+                    spaceBetween: cards.length < 3 ? 75 : 20,
+                  },
+          }}
+        >
+          {cards.map((card, index) => (
+            <SwiperSlide key={index} className={styles.swiperSlide}>
+              <div className={styles.inner__content}>
+                <Card key={index} card={card} customClass={customClass} />
+              </div>
+            </SwiperSlide>
+          ))}
+        </Swiper>
+      </div>
     </Container>
   );
 }