+++ components/snippets/Title.js
@@ -1,5 +1,15 @@
-const Title = ({ title }) => {
-  return <h2>{title}</h2>;
+import styles from '@/styles/Title.module.scss';
+
+const Title = ({ title, customClass }) => {
+  return (
+    <h2
+      className={
+        customClass ? `${styles.title} ${styles[customClass]}` : styles.title
+      }
+    >
+      {title}
+    </h2>
+  );
 };
 
 export default Title;